requirements.txt
aiohttp
discord.py
humanize
jishaku
lavalink
openai
pymongo
quart
Quart_Discord
requests_oauthlib
